+++ superset/assets/src/explore/components/ControlPanelsContainer.jsx
@@ -6,7 +6,7 @@ import { connect } from 'react-redux';
 import { Alert, Tab, Tabs } from 'react-bootstrap';
 import { t } from '@superset-ui/translation';
 
-import visTypes, { sectionsToRender } from '../visTypes';
+import controlPanelConfigs, { sectionsToRender } from '../controlPanelConfigs';
 import ControlPanelSection from './ControlPanelSection';
 import ControlRow from './ControlRow';
 import Control from './Control';
@@ -40,7 +40,7 @@ class ControlPanelsContainer extends React.Component {
     let mapF = controls[controlName].mapStateToProps;
 
     // Looking to find mapStateToProps override for this viz type
-    const controlOverrides = visTypes[this.props.controls.viz_type.value].controlOverrides || {};
+    const controlOverrides = controlPanelConfigs[this.props.controls.viz_type.value].controlOverrides || {};
     if (controlOverrides[controlName] && controlOverrides[controlName].mapStateToProps) {
       mapF = controlOverrides[controlName].mapStateToProps;
     }
