+++ superset/assets/src/explore/controlPanelConfigs.js
